requirements.txt
FastAPI
anyio
celery
loguru
panoptes-utils[config]
pigpio
pydantic>=1.9.0
requests
uvicorn[standard]
